Fix: check for URLs including trailing slash (#13)
Although this is not really relevant in our case, LGTM rightfully
indicates that these 'startswith' can still have URLs like
'https://api.github.com.fake.domain', and validate. We of course
not as much use the validation to check if it really comes from
github.com, merely to ensure that the next statement is valid.
Nevertheless, it is a sane change to make.