Add: integrate Limnoria into the IRC bot, to extend its capabilities
Limnoria is heavily stripped down in what it can do, with most
plugins disabled. The plugins that are enabled, are stripped of
many commands. In short, we want to prevent a Skynet-wannabe.

Limnoria is the successor of supybot, a 15 year old Python
application. They made it compatible with supybot, and as such,
a lot of anti-patterns and weird constructs are in place. We did
our best to minimize what you see of this with this commit, but a
few hacks and patches were required in order to get this to work in
a way that is usable to us.