Fix: flush database-files every 2 minutes
This does increase I/O, but also reduces the lost in data on a
new deployment. It is a balancing act, and 2 minutes sounds like
a reasonable number.
The Seen database of 13 years running supybot is ~1MB. We get
~50KB/s allocated I/O. That means we use 1/6th of the allocation
purely on writing the databases. That sounds like a fine
balance.